# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * VPN Pro API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type VpnServerType = typeof VpnServerType[keyof typeof VpnServerType];


export const VpnServerType = {
  v2ray: 'v2ray',
  proxy: 'proxy',
  socks5: 'socks5',
  shadowsocks: 'shadowsocks',
  trojan: 'trojan',
  hysteria2: 'hysteria2',
  vless: 'vless',
  custom: 'custom',
} as const;

export interface VpnServer {
  id: string;
  name: string;
  type: VpnServerType;
  /** Raw config string (vmess://, ip:port@user:pass, etc.) */
  config: string;
  country: string;
  flag: string;
  ping?: number | null;
  isActive: boolean;
  note?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type CreateServerInputType = typeof CreateServerInputType[keyof typeof CreateServerInputType];


export const CreateServerInputType = {
  v2ray: 'v2ray',
  proxy: 'proxy',
  socks5: 'socks5',
  shadowsocks: 'shadowsocks',
  trojan: 'trojan',
  hysteria2: 'hysteria2',
  vless: 'vless',
  custom: 'custom',
} as const;

export interface CreateServerInput {
  name: string;
  type: CreateServerInputType;
  config: string;
  country: string;
  flag: string;
  note?: string;
}

export type UpdateServerInputType = typeof UpdateServerInputType[keyof typeof UpdateServerInputType];


export const UpdateServerInputType = {
  v2ray: 'v2ray',
  proxy: 'proxy',
  socks5: 'socks5',
  shadowsocks: 'shadowsocks',
  trojan: 'trojan',
  hysteria2: 'hysteria2',
  vless: 'vless',
  custom: 'custom',
} as const;

export interface UpdateServerInput {
  name?: string;
  type?: UpdateServerInputType;
  config?: string;
  country?: string;
  flag?: string;
  note?: string;
}

export interface ActiveServerResponse {
  server: VpnServer | null;
}

export interface ErrorResponse {
  error: string;
}