# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: VPN Pro API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: servers
    description: VPN server management
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /servers:
    get:
      operationId: listServers
      tags: [servers]
      summary: List all VPN servers
      responses:
        "200":
          description: List of servers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VpnServer"

    post:
      operationId: createServer
      tags: [servers]
      summary: Add a new VPN server
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateServerInput"
      responses:
        "201":
          description: Server created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VpnServer"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /servers/{id}:
    get:
      operationId: getServer
      tags: [servers]
      summary: Get a single VPN server
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Server details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VpnServer"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      operationId: updateServer
      tags: [servers]
      summary: Update a VPN server
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateServerInput"
      responses:
        "200":
          description: Server updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VpnServer"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      operationId: deleteServer
      tags: [servers]
      summary: Delete a VPN server
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /servers/{id}/activate:
    post:
      operationId: activateServer
      tags: [servers]
      summary: Set server as the active connection
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Server activated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VpnServer"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /servers/active:
    get:
      operationId: getActiveServer
      tags: [servers]
      summary: Get the currently active server
      responses:
        "200":
          description: Active server or null
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActiveServerResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    VpnServer:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
          enum: [v2ray, proxy, socks5, shadowsocks, trojan, hysteria2, vless, custom]
        config:
          type: string
          description: Raw config string (vmess://, ip:port@user:pass, etc.)
        country:
          type: string
        flag:
          type: string
        ping:
          type: number
          nullable: true
        isActive:
          type: boolean
        note:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - type
        - config
        - country
        - flag
        - isActive
        - createdAt
        - updatedAt

    CreateServerInput:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [v2ray, proxy, socks5, shadowsocks, trojan, hysteria2, vless, custom]
        config:
          type: string
        country:
          type: string
        flag:
          type: string
        note:
          type: string
      required:
        - name
        - type
        - config
        - country
        - flag

    UpdateServerInput:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [v2ray, proxy, socks5, shadowsocks, trojan, hysteria2, vless, custom]
        config:
          type: string
        country:
          type: string
        flag:
          type: string
        note:
          type: string

    ActiveServerResponse:
      type: object
      properties:
        server:
          $ref: "#/components/schemas/VpnServer"
          nullable: true
      required:
        - server

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error