# lib/api-zod/src/generated/types/activeServerResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * VPN Pro API specification
 * OpenAPI spec version: 0.1.0
 */
import type { VpnServer } from './vpnServer';

export interface ActiveServerResponse {
  server: VpnServer | null;
}