# lib/api-zod/src/generated/types/createServerInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * VPN Pro API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateServerInputType } from './createServerInputType';

export interface CreateServerInput {
  name: string;
  type: CreateServerInputType;
  config: string;
  country: string;
  flag: string;
  note?: string;
}