# lib/api-zod/src/generated/types/updateServerInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * VPN Pro API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateServerInputType } from './updateServerInputType';

export interface UpdateServerInput {
  name?: string;
  type?: UpdateServerInputType;
  config?: string;
  country?: string;
  flag?: string;
  note?: string;
}