# lib/api-zod/src/generated/types/vpnServer.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * VPN Pro API specification
 * OpenAPI spec version: 0.1.0
 */
import type { VpnServerType } from './vpnServerType';

export interface VpnServer {
  id: string;
  name: string;
  type: VpnServerType;
  /** Raw config string (vmess://, ip:port@user:pass, etc.) */
  config: string;
  country: string;
  flag: string;
  ping?: number | null;
  isActive: boolean;
  note?: string | null;
  createdAt: string;
  updatedAt: string;
}